category images resolver moved to packages
