+++ packages/frontend-api/src/Resources/config/services.yaml
@@ -106,6 +106,11 @@ services:
             $imagesBatchLoader: '@images_batch_loader'
             $firstImageBatchLoader: '@first_image_batch_loader'
 
+    Shopsys\FrontendApiBundle\Model\Resolver\Image\CategoryOrReadyCategorySeoMixImagesQuery:
+        arguments:
+            $imagesBatchLoader: '@images_batch_loader'
+            $firstImageBatchLoader: '@first_image_batch_loader'
+
     Shopsys\FrontendApiBundle\Model\Resolver\Image\ImagesQuery:
         arguments:
             $imagesBatchLoader: '@images_batch_loader'
